requirements.txt
Flask
chromadb
google-generativeai
gTTS
httpx
pypdf
python-dotenv
